--- packages/lisk-p2p/src/blockchain-p2p/index.ts
@@ -1,16 +1,0 @@
-/*
- * Copyright © 2018 Lisk Foundation
- *
- * See the LICENSE file at the top-level directory of this distribution
- * for licensing information.
- *
- * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
- * no part of this software, including this file, may be copied, modified,
- * propagated, or distributed except according to the terms contained in the
- * LICENSE file.
- *
- * Removal or modification of this copyright notice is prohibited.
- *
- */
-export { BlockchainP2P, IPeerReturnType, IPeerOptions } from './p2p';
-export { IPeerConfig, Peer } from './peer';
